+++ components/dropdown/dropdown-button.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 
 import Button from '../button';
 import type { ButtonHTMLType, ButtonProps } from '../button';
-import type { ButtonGroupProps } from '../button/button-group';
+import type { ButtonGroupProps } from '../button/ButtonGroup';
 import { ConfigContext } from '../config-provider';
 import Space from '../space';
 import { useCompactItemContext } from '../space/Compact';
